requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
twilio==8.5.0
python-dotenv==1.0.0
Flask==2.3.3
gunicorn==21.2.0